chore: Refactor OpenTelemetry configuration and dependencies
